requirements.txt
absl-py==2.4.0
alabaster==1.0.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.4
altair==6.0.0
amqp==5.3.1
annotated-types==0.7.0
astroid==4.0.4
attrs==25.4.0
babel==2.18.0
billiard==4.2.4
black==26.1.0
blinker==1.9.0
cachetools==6.2.6
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmake==4.2.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.4
cycler==0.12.1
Cython==3.2.4
dill==0.4.1
docutils==0.22.4
easydict==1.13
faiss-cpu==1.13.2
filelock==3.24.3
filterpy==1.4.5
flake8==7.3.0
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
iniconfig==2.3.0
insightface==0.7.3
isort==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kombu==5.6.2
lapx==0.9.4
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mediapipe==0.10.32
ml_dtypes==0.5.4
mpmath==1.3.0
mypy_extensions==1.1.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.4.2
onnx==1.20.1
onnxruntime==1.24.2
onnxruntime-gpu==1.24.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
prettytable==3.17.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.5
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
requests==2.32.5
roman-numerals==4.1.0
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
simsimd==6.5.13
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
sounddevice==0.5.5
Sphinx==9.1.0
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.47
streamlit==1.54.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.2.24
toml==0.10.2
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ultralytics==8.4.17
ultralytics-thop==2.0.18
urllib3==2.6.3
vine==5.1.0
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
